Affirmation Report

affirm_report_gt() returns styled gt table summarizing results of affirmation session.

affirm_report_excel() returns excel file with one sheet per affirmation (excluding those with no errors)

affirm_report_raw_data() returns raw data used to generate summary in affirm_report_gt()

Usage

affirm_report_gt()

affirm_report_excel(
  file,
  affirmation_name = "{data_frames}{id}",
  overwrite = TRUE,
  previous_file = NULL
)

affirm_report_raw_data()

Arguments

file

A file path to save the xlsx file

affirmation_name

A string for affirmation names; the item name in curly brackets is replaced with the item value (see glue::glue). Item names accepted include: id, label, priority, data_frames, columns, error_n, total_n. Defaults to "{data_frames}{id}".

overwrite

Overwrite existing file (Defaults to TRUE as with write.table)

previous_file

A string of the file path to the previous affirmation Excel workbook containing assigned_to, status, and comment fields that need to be carried forward to this report

Value

gt table, Excel file, or a df of raw data

Updating Previous Excel Affirm Reports with affirm_report_excel()

As of version 0.2.1, Excel affirm reports can now be updated with data from previous reports. This feature allows the assigned_to, status, and comment columns in newly created Excel affirm reports to be populated with values from a previous Excel affirm report.

To successfully update a newly created Excel affirm report, three conditions must be met:

Previous Summary Sheet: The previous affirm report used for updating must contain a tab named 'summary' as the first tab in the Excel workbook. Any additional tabs in the previous report that are not affirmations will be omitted from the updated affirm report.

Matching Affirmation Columns: When updating a previous affirmation, columns cannot be dropped from the new affirmation. All columns from the previous report are required for merging old values with new values in the updated report. If an affirmation tab has missing columns compared to the previous report, the update will fail. However, new affirmations can include additional columns and still be updated successfully, as long as all columns from the previous report are present.

Unique Affirmation Rows: Both the previous and newly created affirmations must not contain any duplicate rows. This requirement ensures that data from the previous affirmation can be accurately merged with the new affirmations through join operations.

Examples

affirm_init(replace = TRUE)
#> ✔ We're ready to make data affirmations...

dplyr::as_tibble(mtcars) |>
 affirm_true(
   label = "No. cylinders must be 4, 6, or 8",
   condition = cyl %in% c(4, 6, 8)
 ) |>
 affirm_true(
    label = "MPG should be less than 33",
    condition = mpg < 33
 )
#> • No. cylinders must be 4, 6, or 8
#>   0 issues identified.
#> • MPG should be less than 33
#>   1 issue identified.
#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # ℹ 22 more rows

gt_report <- affirm_report_gt()

affirm_close()