Affirm True

Use this function to affirm an expression is true.

Usage

affirm_true(
  data,
  label,
  condition,
  id = NA_integer_,
  priority = NA_integer_,
  data_frames = NA_character_,
  columns = NA_character_,
  report_listing = NULL,
  data_action = NULL,
  error = getOption("affirm.error", default = FALSE)
)

Arguments

data

a data frame

label

a string used to describe the affirmation

condition

expression to check that evaluates to a logical vector, e.g. cyl %in% c(4, 6, 8). Use the dot (.) to reference the passed data frame. If condition results in a missing value, it is interpreted as FALSE.

id, priority, data_frames, columns

Optional additional information that will be passed to affirmation report.

id must be an integer, e.g. id = 1L

priority must be an integer, e.g. priority = 1L

data_frames string of data frame names used in affirmation, e.g. data_frames = "RAND, DM"

columns string of column names used in affirmation. default is all.vars(condition)

report_listing

an expression selecting/filtering rows from data= to return in the issue listing report. The default is to return the result from create_report_listing(), which are the rows that do not met in condition= and columns included in the condition= expression along with any columns set in option('affirm.id_cols'). The 'affirm.id_cols' option must be a character vector of column names, where columns will be selected with select(any_of(getOption('affirm.id_cols'))).

data_action

this expression is executed at the end of the function call when supplied.

Default is NULL, and the passed data frame in data= is returned unaltered.

Perhaps you'll need to remove problematic rows: data_action = filter(., !(!!condition))

error

Logical indicating whether to throw an error when condition is not met. Default is FALSE.

Value

data frame

Details

When passing expressions to arguments report_listing= and data_action=, there are a few things to keep in mind.

The expression passed in condition= can be used, but note that it has been captured as an expression inside the function. This means that to use it, you'll need to use !! (bang-bang) to pass it inside a function.

In addition to being able to use the condition= expression, you can simplify your code somewhat by referring to lgl_condition, which is an evaluated logical vector of the condition= expression.

See also

Examples

affirm_init(replace = TRUE)
#> ✔ We're ready to make data affirmations...

as_tibble(mtcars) |>
 affirm_true(
   label = "No. cylinders must be 4, 6, or 8",
   condition = cyl %in% c(4, 6, 8)
 )
#> • No. cylinders must be 4, 6, or 8
#>   0 issues identified.
#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # ℹ 22 more rows

affirm_close()